requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.6.4
pydantic-settings==2.2.1
redis==5.0.3
rq==1.15.1
httpx==0.27.0
ccxt==4.3.80
pandas==2.2.2
numpy==1.26.4
prophet==1.1.5
python-dotenv==1.0.1
prometheus-client==0.20.0
requests==2.31.0
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==5.0.0
hypothesis==6.100.1
locust==2.22.0
mypy==1.8.0
ruff==0.3.4
email-validator==2.2.0
